requirements.txt
streamlit==1.28.1
openai==1.3.7
python-dotenv==1.0.0
neo4j==5.14.1
PyPDF2==3.0.1
pdfplumber==0.10.3